requirements.txt
python-telegram-bot==20.7
requests==2.31.0
pytz==2024.1
python-dotenv==1.0.0
timezonefinder==6.2.0